PsycheGraphFrontend-main/backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.1
numpy==1.26.4
librosa==0.10.1
soundfile==0.12.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-jose==3.3.0
boto3==1.34.41
openai==1.12.0
python-multipart
pydub==0.25.1
greenlet>=3.0.0
sqlalchemy==2.0.42
asyncpg==0.30.0
pydantic-settings==2.7.1
passlib[bcrypt]==1.7.4
structlog==23.2.0
alembic==1.12.1
httpx==0.24.1
pytest==8.3.4
pytest-asyncio==0.22.0
email-validator>=2.0.0
